# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 极简栈创造 AI 生成接口
  description: |
    涉及 AI 适配器层的对外接口规范，包括模型绑定、生成任务、Adapter 健康检查。
  version: 1.0.0
  contact:
    name: Core · 后端工程师
servers:
  - url: /api/v1
    description: 本地开发服务
paths:
  /projects:
    get:
      summary: 列出项目
      description: 返回当前用户的项目列表，支持状态筛选、标题搜索、排序和分页。
      operationId: listProjects
      tags:
        - Projects
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - name: search
          in: query
          schema:
            type: string
          description: 标题关键字搜索
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - updated_at_asc
              - updated_at_desc
            default: updated_at_desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/ProjectList'
    post:
      summary: 创建项目
      description: 创建新项目，初始状态为 draft，不触发 AI 生成。
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 获取项目详情
      operationId: getProject
      tags:
        - Projects
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: 更新项目
      operationId: updateProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 删除项目
      operationId: deleteProject
      tags:
        - Projects
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
...
  /projects/{projectId}/locations:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 列出项目场景卡片
      description: 返回指定项目下的所有场景（location）卡片。
      operationId: listLocations
      tags:
        - Locations
      responses:
        '200':
          description: 场景卡片列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/LocationList'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: 更新场景卡片
      description: 更新场景字段，自动触发版本快照。
      operationId: updateLocation
      tags:
        - Locations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}/generate-base:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 生成基准候选图
      description: 基于场景字段生成 3 张基准候选图，返回候选 URL 与 seed，不持久化。
      operationId: generateBaseCandidates
      tags:
        - Locations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateBaseCandidatesRequest'
      responses:
        '200':
          description: 候选图列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BaseCandidate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}/confirm-base:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 确认基准图
      description: 确认选中的基准候选图，持久化 base_seed 与 base_image_url，触发版本快照。
      operationId: confirmBaseImage
      tags:
        - Locations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmBaseRequest'
      responses:
        '200':
          description: 确认成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}/generate-variant:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 生成变体图
      description: 基于已确认的 base_seed 与固定偏移生成指定变体图，返回 URL 与 seed，不持久化。
      operationId: generateLocationVariant
      tags:
        - Locations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateVariantRequest'
      responses:
        '200':
          description: 变体图结果
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/VariantResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}/confirm-variant:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 确认变体图
      description: 确认变体图并持久化到 variants 字典，key 为 `{time_of_day}-{weather}`，触发版本快照。
      operationId: confirmLocationVariant
      tags:
        - Locations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmVariantRequest'
      responses:
        '200':
          description: 确认成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}/versions:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 获取场景历史版本
      description: 返回场景卡片的版本快照列表（不含完整 content）。
      operationId: listLocationVersions
      tags:
        - Locations
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: 版本历史
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/SnapshotHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/locations/{locationId}/rollback:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 回滚场景到指定版本
      description: 基于历史版本快照恢复场景卡片数据。
      operationId: rollbackLocation
      tags:
        - Locations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackLocationRequest'
      responses:
        '200':
          description: 回滚成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    ProjectStatus:
      type: string
      enum:
        - draft
        - outlining
        - asset_prep
        - producing
        - completed
    ProjectGenre:
      type: string
      enum:
        - urban_romance
        - ancient_costume
        - suspense
        - comedy
        - sci_fi
        - other
    ProjectDurationGoal:
      type: string
      enum:
        - 3min
        - 5min
        - 10min
    ProjectStyleTag:
      type: string
      enum:
        - realistic
        - comic
        - cyberpunk
        - chinese_style
        - fresh
        - dark
    ProjectMeta:
      type: object
      required:
        - title
        - description
        - genre
        - style_tags
      properties:
        title:
          type: string
          maxLength: 50
          description: 项目标题
        description:
          type: string
          maxLength: 1000
          description: 创意描述
        genre:
          $ref: '#/components/schemas/ProjectGenre'
        target_episodes:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
          description: 目标集数，留空表示由 AI 后续建议
        duration_goal:
          $ref: '#/components/schemas/ProjectDurationGoal'
          nullable: true
          description: 时长目标
        style_tags:
          type: array
          items:
            $ref: '#/components/schemas/ProjectStyleTag'
          description: 风格标签
        notes:
          type: string
          maxLength: 2000
          nullable: true
          description: 创作备注
    Project:
      type: object
      required:
        - id
        - user_id
        - status
        - meta
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        team_id:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        meta:
          $ref: '#/components/schemas/ProjectMeta'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectList:
      type: object
      required:
        - total
        - projects
      properties:
        total:
          type: integer
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    CreateProjectRequest:
      type: object
      required:
        - meta
      properties:
        meta:
          $ref: '#/components/schemas/ProjectMeta'
    UpdateProjectRequest:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/ProjectMeta'
        status:
          $ref: '#/components/schemas/ProjectStatus'
...
    SnapshotSource:
      type: string
      enum:
        - ai_generated
        - user_edited
        - ai_regenerated
        - locked
      description: 版本快照来源
    AIModelInfo:
      type: object
      required:
        - provider
        - model
      properties:
        provider:
          type: string
          example: openai
        model:
          type: string
          example: gpt-4o
    SnapshotMeta:
      type: object
      required:
        - id
        - versionId
        - versionNumber
        - source
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        versionId:
          type: string
          example: v3
          pattern: '^v\d+$'
        versionNumber:
          type: integer
          example: 3
        source:
          $ref: '#/components/schemas/SnapshotSource'
        editedBy:
          type: string
          nullable: true
          description: 用户编辑时的用户 ID
        aiModel:
          $ref: '#/components/schemas/AIModelInfo'
          nullable: true
          description: AI 生成时记录的模型配置
        promptOverride:
          type: string
          nullable: true
          description: 用户给 AI 的额外指令
        parentVersionNumber:
          type: integer
          nullable: true
          description: 父版本号
        createdAt:
          type: string
          format: date-time
...
    SnapshotHistory:
      type: object
      required:
        - total
        - versions
      properties:
        total:
          type: integer
        versions:
          type: array
          items:
            $ref: '#/components/schemas/SnapshotMeta'
...
    LocationStatus:
      type: string
      enum:
        - draft
        - confirmed
      description: 场景卡片状态
    LocationVariant:
      type: object
      required:
        - image_url
        - prompt
        - seed
      properties:
        image_url:
          type: string
          format: uri
        prompt:
          type: string
        seed:
          type: integer
    Location:
      type: object
      required:
        - id
        - project_id
        - name
        - status
        - key_props
        - variants
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        frequency:
          type: string
          nullable: true
        space_type:
          type: string
          nullable: true
        style:
          type: string
          nullable: true
        color_tone:
          type: string
          nullable: true
        lighting_type:
          type: string
          nullable: true
        key_props:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/LocationStatus'
        base_seed:
          type: integer
          nullable: true
        base_image_url:
          type: string
          format: uri
          nullable: true
        variants:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/LocationVariant'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LocationList:
      type: object
      required:
        - total
        - locations
      properties:
        total:
          type: integer
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
    UpdateLocationRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 2000
          nullable: true
        frequency:
          type: string
          maxLength: 50
          nullable: true
        space_type:
          type: string
          maxLength: 50
          nullable: true
        style:
          type: string
          maxLength: 200
          nullable: true
        color_tone:
          type: string
          maxLength: 100
          nullable: true
        lighting_type:
          type: string
          maxLength: 100
          nullable: true
        key_props:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/LocationStatus'
    BaseCandidate:
      type: object
      required:
        - url
        - seed
        - prompt
      properties:
        url:
          type: string
          format: uri
        seed:
          type: integer
        prompt:
          type: string
    GenerateBaseCandidatesRequest:
      type: object
      properties:
        seed:
          type: integer
          description: 固定种子，不传则根据 locationId 计算
    ConfirmBaseRequest:
      type: object
      required:
        - candidate
      properties:
        candidate:
          $ref: '#/components/schemas/BaseCandidate'
    VariantResult:
      type: object
      required:
        - url
        - seed
        - prompt
      properties:
        url:
          type: string
          format: uri
        seed:
          type: integer
        prompt:
          type: string
    GenerateVariantRequest:
      type: object
      required:
        - time_of_day
        - weather
      properties:
        time_of_day:
          type: string
        weather:
          type: string
    ConfirmVariantRequest:
      type: object
      required:
        - time_of_day
        - weather
        - variant
      properties:
        time_of_day:
          type: string
        weather:
          type: string
        variant:
          $ref: '#/components/schemas/VariantResult'
    RollbackLocationRequest:
      type: object
      required:
        - version_id
      properties:
        version_id:
          type: string
          pattern: '^v\d+$'
          description: 目标版本号
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 统一错误码
          example: AI_TIMEOUT
        message:
          type: string
          description: 错误信息（不暴露堆栈）
        retryable:
          type: boolean
          description: 是否为可用性错误，可触发 fallback
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'